# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Antojados API
  version: 1.0.0
  description: API REST de AntojadosMX - Auth, Feed, Places, Posts, Social, Biz, Rewards, Media, Sync, Analytics, Equipo, GT Panel.
servers:
  - url: https://api.antojadosmx.mx
    description: Produccion (Cloudflare HTTPS)
  - url: http://185.187.235.253:8010
    description: Contabo (IP directa - fallback)
...
paths:
...
  /api/v1/antojados/places/{place_id}:
    get:
      tags: [Places]
      summary: Obtener lugar
      parameters: [{in: path, name: place_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
    patch:
      tags: [Places]
      summary: Actualizar lugar
      parameters: [{in: path, name: place_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/places/{place_id}/posts:
    get:
      tags: [Places]
      summary: Posts de un lugar
      parameters: [{in: path, name: place_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/places/{place_id}/ratings-summary:
    get:
      tags: [Places]
      summary: Resumen de ratings de un lugar
      parameters: [{in: path, name: place_id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/tenants/{id}:
    get:
      tags: [GT Tenants]
      summary: Detalle de tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/activate:
    post:
      tags: [GT Tenants]
      summary: Activar tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/suspend:
    post:
      tags: [GT Tenants]
      summary: Suspender tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/reactivate:
    post:
      tags: [GT Tenants]
      summary: Reactivar tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/notifications/{id}/read:
    patch:
      tags: [GT Notifications]
      summary: Marcar notificacion como leida
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/notification-sequences/{id}/assign:
    post:
      tags: [GT Notifications]
      summary: Asignar secuencia de notificacion
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/settings/{key}:
    patch:
      tags: [GT Ops]
      summary: Actualizar setting GT
      parameters: [{in: path, name: key, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
...
  /api/v1/antojados/gt/tenants/{id}/audit:
    get:
      tags: [GT Ops]
      summary: Audit log de tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/suspensions:
    get:
      tags: [GT Ops]
      summary: Historial de suspensiones de tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/economic-snapshot:
    get:
      tags: [GT Ops]
      summary: Snapshot economico del tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/economic-snapshot/sync:
    post:
      tags: [GT Ops]
      summary: Sincronizar snapshot economico
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/economic-event:
    post:
      tags: [GT Ops]
      summary: Registrar evento economico del tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "201": { description: Created }
  /api/v1/antojados/gt/tenants/{id}/packages:
    get:
      tags: [GT Ops]
      summary: Paquetes del tenant
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        "200": { description: OK }
  /api/v1/antojados/gt/tenants/{id}/packages/{code}:
    post:
      tags: [GT Ops]
      summary: Activar paquete para tenant
      parameters:
        - {in: path, name: id, required: true, schema: {type: string}}
        - {in: path, name: code, required: true, schema: {type: string}}
      responses:
        "200": { description: OK }
...